Fix tests for 64-bit.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/apple-local-200502-branch@109798 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.apple/bitreverse-15.c
@@ -38,7 +38,7 @@ typedef struct _S4
         unsigned short sbf2 : 8;
         unsigned short sbf3 : 8;
         unsigned short sbf4 : 8;
-        unsigned long l;
+        unsigned int l;
 } S4;
 #pragma pack(pop)
